pyproject.toml
[project]
name = "benten"
version = "0.1.0"
description = "Daily countdown (100→67) Telegram user bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "telethon",
    "pyyaml",
    "python-dotenv",
    "tzdata",
]

[dependency-groups]
dev = []
